refactor(identity): keep session model out of outbound port (#770)
