+++ src/main/java/de/caritas/cob/userservice/api/model/identity/IdentitySession.java
@@ -1,4 +1,4 @@
-package de.caritas.cob.userservice.api.port.out.identity;
+package de.caritas.cob.userservice.api.model.identity;
 
 /** Provider-neutral authenticated session returned by an identity-provider exchange. */
 public record IdentitySession(
